requirements.txt
openai
python-dotenv
unittest